Multivalue u64 field / i64 field.

+++ src/postings/mod.rs
@@ -25,7 +25,7 @@ pub use self::segment_postings::{BlockSegmentPostings, SegmentPostings};
 
 pub use common::HasLen;
 
-pub(crate) type UnorderedTermId = usize;
+pub(crate) type UnorderedTermId = u64;
 
 #[allow(enum_variant_names)]
 pub(crate) enum FreqReadingOption {
